Changed a few simple things
Fixed a simple bug with the FPS not being displayed properly.
Renamed some variables to be more consistent.
Refactored some code for better logic.

+++ mPULSE refactored/mPULSE/mPULSE.cpp
@@ -14,7 +14,7 @@ int frameRate = 60;
 Game game;
 XboxController* players[MAX_PLAYERS];
 Clock gameClock(frameRate);
-unsigned int deltaTime = 0;
+float deltaTime = 0;
 SDL_Event KeyboardMouseState;    
 XINPUT_STATE XboxState;
